[DT] Set encodings if `iree.opt.data-tiling` unit attribute is attached.
The revision moves the filter to AnnotateDataTilingHints pass that adds
the unit attribute to target gemms.

If any operation has the data-tiling hint, it does nothing. It allow
users to perform data-tiling selectively in preprocessing. E.g., user
can run a transform dialect script that matches linalg ops and attaches
the unit attribute.

E.g., the below transform script only attaches `iree.opt.data-tiling` to
the first matmul.

```
func.func @matmuls(%lhs: tensor<?x?xi8>, %rhs: tensor<?x?xi8>, %acc: tensor<?x?xi32>) -> (tensor<?x?xi32>, tensor<?x?xi32>) {
  %res0 = linalg.matmul_transpose_b
        ins(%lhs, %lhs : tensor<?x?xi8>, tensor<?x?xi8>)
        outs(%acc : tensor<?x?xi32>) -> tensor<?x?xi32>
  %res1 = linalg.matmul_transpose_b
        ins(%lhs, %rhs : tensor<?x?xi8>, tensor<?x?xi8>)
        outs(%acc : tensor<?x?xi32>) -> tensor<?x?xi32>
  return %res0, %res1 : tensor<?x?xi32>, tensor<?x?xi32>
}

module attributes {transform.with_named_sequence} {
  transform.named_sequence @match_matmul_repeated_operand(%arg0: !transform.any_op {transform.readonly}) -> !transform.any_op {
    %inputs, %outputs = transform.iree.match.cast_compatible_dag_from_root %arg0 {
    ^bb0(%arg1: tensor<?x?xi8>, %arg2: tensor<?x?xi32>):
      %1 = linalg.matmul_transpose_b
          ins(%arg1, %arg1 : tensor<?x?xi8>, tensor<?x?xi8>)
          outs(%arg2 : tensor<?x?xi32>) -> tensor<?x?xi32>
    } : (!transform.any_op) -> (!transform.any_value, !transform.any_value)
    transform.yield %arg0 : !transform.any_op
  }

  transform.named_sequence @annotate(%generic: !transform.any_op {transform.readonly}) {
    transform.annotate %generic "iree.opt.data-tiling" : !transform.any_op
    transform.yield
  }

  transform.named_sequence @__transform_main(%module: !transform.any_op) {
    %func = transform.structured.match ops{["func.func"]} in %module : (!transform.any_op) -> !transform.any_op
    transform.foreach_match in %module
        @match_matmul_repeated_operand -> @annotate
      : (!transform.any_op) -> (!transform.any_op)
    transform.yield
  }
}
```
