+++ compiler/src/iree/compiler/DispatchCreation/BUILD.bazel
@@ -20,6 +20,7 @@ package(
 iree_compiler_cc_library(
     name = "DispatchCreation",
     srcs = [
+        "AnnotateDataTilingHints.cpp",
         "BitcastUnsupportedElementTypes.cpp",
         "BubbleUpExpandShapes.cpp",
         "CloneProducersIntoDispatchRegions.cpp",
