+++ compiler/src/iree/compiler/DispatchCreation/Passes.cpp
@@ -271,6 +271,7 @@ addDispatchRegionCreationPasses(OpPassManager &passManager,
           options.cseConstants = false;
           return IREE::Flow::createCanonicalizePass(options);
         })
+        .addPass(createAnnotateDataTilingHintsPass)
         // Set encodings on all eligible ops. All ops should be in compiler
         // formed dispatch regions, so encodings will be placed inside of the
         // dispatch regions with the data-tiled op.
