+++ compiler/src/iree/compiler/DispatchCreation/test/BUILD.bazel
@@ -16,6 +16,7 @@ iree_lit_test_suite(
     name = "lit",
     srcs = enforce_glob(
         [
+            "annotate_data_tiling_hints.mlir",
             "bitcast_unsupported_element_types.mlir",
             "clone_producers_into_dispatch_regions.mlir",
             "collapse_dimensions.mlir",
