+++ compiler/src/iree/compiler/GlobalOptimization/Passes.cpp
@@ -181,12 +181,12 @@ void buildGlobalOptimizationPassPipeline(
 
   // Enable data tiling after they are in a canonical form.
   if (transformOptions.options.dataTiling) {
-    FunctionLikeNest(mainPassManager).addPass([&]() {
-      return DispatchCreation::createSetEncodingPass(
-          DispatchCreation::SetEncodingPassOptions{clSetEncodingStrategy});
-    });
-    // TODO(hanchung): Make data-tiling passes be FunctionOpInterface pass, so
-    // we can use `FunctionLikNest` here.
+    FunctionLikeNest(mainPassManager)
+        .addPass(DispatchCreation::createAnnotateDataTilingHintsPass)
+        .addPass([&]() {
+          return DispatchCreation::createSetEncodingPass(
+              DispatchCreation::SetEncodingPassOptions{clSetEncodingStrategy});
+        });
     if (clEnableEarlyMaterialization) {
       mainPassManager.addPass(createMaterializeHomogeneousEncodingsPass());
     }
